feat: add repository for ward
